Smart accumulation
The idea behind this kind of accumulation is to see the actual charts
for every shown week.  For example, when you move the mouse to see
artist charts for a week that happened to be a month ago, you'll see the
same charts as if you visited Last.fm a month ago and requested the top
artist charts.

It requires twice more weekly charts for every timeframe except
'overall'.